requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0